Dirty rects removed, flicker added
    I attempted this time to reshape the program by brute force, substituting
the faster dirty rect animation for full screen redraws. It behaves as
expected for a single circle, but as soon a second circle appears, the screen
flickers terribly, like an analog antenna TV that's out-of-tune. The good news
is, the circles still appear as drawn. Horribly inefficient, but it forces me
to rethink the problem.
    I may still be able to use dirty rect animation, although I'll have to
detect circles in the area of the one being drawn, and redraw them as well. I
can use the rects of each circle and some simple comparisons to see who's
around.